Instance files expanding

+++ src/main/kotlin/com/cognifide/gradle/aem/instance/AemLocalHandler.kt
@@ -1,6 +1,8 @@
 package com.cognifide.gradle.aem.instance
 
 import com.cognifide.gradle.aem.AemConfig
+import com.cognifide.gradle.aem.AemException
+import com.cognifide.gradle.aem.AemPlugin
 import com.cognifide.gradle.aem.internal.FileOperations
 import com.cognifide.gradle.aem.internal.Patterns
 import com.cognifide.gradle.aem.internal.PropertyParser
@@ -9,9 +11,10 @@ import org.gradle.api.Project
 import org.gradle.api.logging.Logger
 import org.gradle.util.GFileUtils
 import java.io.File
+import java.io.Serializable
 
 /**
- * TODO put each dir into create:destroy outputs / task caching
+ * TODO Manage background AEM process properly (java-service-wrapper or commons-exec?)
  */
 class AemLocalHandler(val base: AemInstance, val project: Project) {
 
@@ -34,39 +37,69 @@ class AemLocalHandler(val base: AemInstance, val project: Project) {
     val license = File(dir, "license.properties")
 
     fun create(files: List<File>) {
-        logger.info("Copying instance files to '${dir.absolutePath}'")
+        logger.info("Creating instance at path '${dir.absolutePath}'")
+
+        val filesDir = File(config.instanceFilesPath)
+
+        logger.info("Copying configured instance files from '${filesDir.absolutePath}'")
+        if (filesDir.exists()) {
+            FileUtils.copyDirectory(filesDir, dir)
+        }
+
+        logger.info("Copying resolved instance files: ${files.map { it.absolutePath }}")
         GFileUtils.mkdirs(dir)
         files.forEach { FileUtils.copyFileToDirectory(it, dir) }
 
-        logger.info("JAR file found: ${jar.absolutePath}, exists: ${jar.exists()}")
-        logger.info("License file found: ${license.absolutePath}, exists: ${jar.exists()}")
+        logger.info("Copying missing instance files (preserving defaults")
+        FileOperations.copyResources(AemPlugin.INSTANCE_FILES_PATH, dir, true)
 
-        // TODO make these files configurable (firstly copy from project), then use resources, then expand
-        FileOperations.copyResources("local-instance", dir, false, { file, input ->
-            if (Patterns.wildcard(file, listOf("**/*.bat", "**/*.sh"))) {
-                val text = input.bufferedReader().use { it.readText() }
-
-                PropertyParser(project).expand(text, mapOf(
-                        "instance" to base,
-                        "jar" to jar,
-                        "license" to license
-                )).byteInputStream()
-            } else {
-                input
-            }
+        logger.info("Expanding instance files")
+        FileOperations.amendFiles(dir, config.instanceFilesExpanded, { source ->
+            PropertyParser(project).expand(source, properties)
         })
+
+        logger.info("Created instance with success")
     }
+
+    fun validate() {
+        if (!jar.exists()) {
+            throw AemException("Instance JAR file not found at path: ${jar.absolutePath}")
+        }
+
+        if (!license.exists()) {
+            throw AemException("License file not found at path: ${license.absolutePath}")
+        }
+    }
+
+    val properties: Map<String, Serializable>
+        get() {
+            return mapOf(
+                    "instance" to base,
+                    "jar" to jar,
+                    "license" to license
+            )
+        }
 
     fun destroy() {
-        dir.deleteRecursively()
+        logger.info("Destroying instance at path '${dir.absolutePath}'")
+
+        if (dir.exists()) {
+            dir.deleteRecursively()
+
+        }
+
+        logger.info("Destroyed instance with success")
     }
 
+
+    // TODO Implement 'up'
     fun up() {
         Runtime.getRuntime().exec("java -jar ${jar.absolutePath}")
     }
 
+    // TODO Implement 'down'
     fun down() {
-        // TODO ...
+        // ...
     }
 
     override fun toString(): String {
